리뷰 추가 API 구현

See merge request s06-bigdata-dist-sub2/S06P22A301!6

+++ backend/src/main/java/com/showing/backend/db/entity/performance/Actor.java
@@ -22,7 +22,7 @@ public class Actor extends BaseTimeEntity implements Serializable {
     private String actorName;
 
     @Column(unique = true)
-    private int playdbId;
+    private Long playdbId;
 
     @Column
     private String actorImage;
